+++ src/main/java/chess/domain/member/Member.java
@@ -1,4 +1,4 @@
-package chess.dao;
+package chess.domain.member;
 
 public class Member {
